Break out samples and bs from GeoEstimatorWidget into separate widget
